feat(server): 新增分类管理功能并优化相关接口
- 新增分类管理相关的 CRUD 接口和页面查询接口
- 实现分类树形结构的构建和分页查询
- 添加每个分类中的文章数量统计接口
- 优化了数据库查询效率，减少了冗余代码

+++ src/main/java/top/luoyuanxiang/thrivex/server/entity/Assistant.java
@@ -56,7 +56,7 @@ public class Assistant extends Model<Assistant> {
      * 是否被启用
      */
     @TableField("is_default")
-    private Byte isDefault;
+    private Integer isDefault;
 
     @Override
     public Serializable pkVal() {
